+++ src/hooks/index.js
@@ -5,6 +5,7 @@ export * from './useArticlePagination'
 export * from './useArticlePreviewTracking'
 export * from './useArticleSummary'
 export * from './useErrorLogging'
+export * from './useExperimentConfig'
 export * from './useHistoryState'
 export * from './useI18n'
 export * from './useNavigation'
